+++ example/android/app/src/main/java/io/flutter/plugins/GeneratedPluginRegistrant.java
@@ -1,6 +1,8 @@
 package io.flutter.plugins;
 
 import io.flutter.plugin.common.PluginRegistry;
+import io.flutter.plugins.flutter_plugin_android_lifecycle.FlutterAndroidLifecyclePlugin;
+import io.flutter.plugins.imagepicker.ImagePickerPlugin;
 import org.basil.launcher_helper.LauncherHelperPlugin;
 import com.baseflow.permissionhandler.PermissionHandlerPlugin;
 
@@ -12,6 +14,8 @@ public final class GeneratedPluginRegistrant {
     if (alreadyRegisteredWith(registry)) {
       return;
     }
+    FlutterAndroidLifecyclePlugin.registerWith(registry.registrarFor("io.flutter.plugins.flutter_plugin_android_lifecycle.FlutterAndroidLifecyclePlugin"));
+    ImagePickerPlugin.registerWith(registry.registrarFor("io.flutter.plugins.imagepicker.ImagePickerPlugin"));
     LauncherHelperPlugin.registerWith(registry.registrarFor("org.basil.launcher_helper.LauncherHelperPlugin"));
     PermissionHandlerPlugin.registerWith(registry.registrarFor("com.baseflow.permissionhandler.PermissionHandlerPlugin"));
   }