+++ lib/launcher_helper.dart
@@ -10,4 +10,4 @@ library launcher_helper;
 
 export 'src/_launcher_helper.dart';
 export 'src/palette_generator.dart';
-export 'src/apps_info.dart';
+export 'src/application_collection.dart';