+++ lib/src/_launcher_helper.dart
@@ -1,27 +1,14 @@
-/// Copyright 2019 Mushaheed Syed. All rights reserved.
-///
-/// Use of this source code is governed by a MIT license that can be
-/// found in the LICENSE file.
-///
-/// --------------------------------------------------------------------------------
-/// Copyright 2017 Ashraff Hathibelagal
-///
-/// Licensed under the Apache License, Version 2.0 (the "License");
-/// you may not use this file except in compliance with the License.
-/// You may obtain a copy of the License at
-///     http://www.apache.org/licenses/LICENSE-2.0
-/// Unless required by applicable law or agreed to in writing, software
-/// distributed under the License is distributed on an "AS IS" BASIS,
-/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-/// See the License for the specific language governing permissions and
-/// limitations under the License.
+// Copyright 2019 Mushaheed Syed. All rights reserved.
+
+// Use of this source code is governed by a MIT license that can be
+// found in the LICENSE file.
 
 import 'dart:async';
 import 'dart:typed_data';
 
 import 'package:flutter/services.dart';
-import 'package:flutter/foundation.dart';
-import 'apps_info.dart';
+import 'package:flutter/widgets.dart';
+import 'application_collection.dart';
 import 'palette_generator.dart';
 
 /// # LauncherHelper
@@ -31,26 +18,45 @@ import 'palette_generator.dart';
 /// ## Available methods/getters:
 ///
 /// - Use [getApplications] to get list of apps installed.
+/// - [getApplicationInfo] to get [Application] for provided `packageName`.
+/// - [doesApplicationExist] to check if application with provided `packageName` exists.
+/// - [getApplicationIcon] returns app icon for provided `packageName`.
+/// - [isApplicationEnabled] returns `true` if application is enabled else returns `false`.
 /// - [launchApp] can launch apps by providing their package name.
 /// - [getWallpaper] returns device wallpaper as [Uint8List].
-/// - [getWallpaperBrightness] returns brightness of wallpaper.
-/// - [getLuminance] calculates approximate luminance/brightness of an image.
-/// - [getBrightnessFrom] calculates brightness of an image from every pixel.
-/// - [wallpaperPalette] generated color palettes of wallpaper using [PaletteGenerator].
+/// - [calculateBrightness] calculates brightness of an image from every pixel.
+/// - [generatePalette] generates color palettes from `Uint8List` image data using [PaletteGenerator].
 class LauncherHelper {
   static const MethodChannel _channel = const MethodChannel('launcher_helper');
 
-  /// Returns a list of apps installed on the user's device
-  @deprecated
-  static Future<List> get getApps async {
-    var data = await _channel.invokeMethod('getAllApps');
-    return data;
-  }
-
-  /// Returns an [ApplicationCollection] object with [Application] of apps installed on the user's device.
+  /// Returns [ApplicationCollection] of [Application]s installed on this device.
   static Future<ApplicationCollection> get getApplications async {
     List data = await _channel.invokeMethod('getAllApps');
     return ApplicationCollection(data);
+  }
+
+  /// Returns [Application] matching with provided `packageName` installed on this device. Throws "No_Such_App_Found" exception if app doesn't exists.
+  static Future<Application> getApplicationInfo(String packageName) async {
+    Map data = await _channel.invokeMethod('getApplicationInfo', {"packageName": packageName});
+    return Application.fromMap(data);
+  }
+
+  /// Returns true if application exists else false if it doesn't exist. Throws "No_Such_App_Found" exception if app doesn't exists.
+  static Future<bool> doesApplicationExist(String packageName) async {
+    bool data = await _channel.invokeMethod('doesAppExist', {"packageName": packageName});
+    return data;
+  }
+
+  /// Returns true if application is enabled else false if it isn't enabled. Throws "No_Such_App_Found" exception if app doesn't exists.
+  static Future<bool> isApplicationEnabled(String packageName) async {
+    bool data = await _channel.invokeMethod('isAppEnabled', {"packageName": packageName});
+    return data;
+  }
+
+  /// Returns application. Throws "No_Such_App_Found" exception if app or app-icon doesn't exists.
+  static Future<bool> getApplicationIcon(String packageName) async {
+    bool data = await _channel.invokeMethod('getIconOfPackage', {"packageName": packageName});
+    return data;
   }
 
   /// Launches an app using its package name
@@ -73,54 +79,6 @@ class LauncherHelper {
     return data;
   }
 
-  /// This gets the brightness of current Wallpaper to determine theme (light or dark). The function returns
-  /// a brightness level between 0 and 255, where 0 = totally black and 255 = totally bright.
-  ///
-  /// `skipPixel` parameter refers to number of pixels to skip while calculating Wallpaper's brightness.
-  /// `skipPixel` defaults to 1 (every pixel is counted) and can't be less than 1.
-  ///
-  /// __Note:__
-  /// - This method needs the READ_EXTERNAL_STORAGE permission on Android Oreo & above.
-  static Future<int> getWallpaperBrightness({int skipPixel = 1}) async {
-    assert(skipPixel > 0, 'skipPixel should have a value greater than 0');
-    debugPrint(
-        "[LauncherHelper] External Storage Access permission might be needed for Android Oreo & above.");
-    int data = await _channel
-        .invokeMethod('getWallpaperBrightness', {'skipPixel': skipPixel});
-    return data;
-  }
-
-  /// This asynchronously calculates luminance for an image using dominant colors from palettes of
-  /// image.
-  ///
-  /// The function returns a [double] representing `luminance` from image data of `Uint8List` type.
-  /// `luminance` with a brightness value between 0 for darkest and 1 for lightest.
-  /// It represents the relative luminance of the color.
-  ///
-  /// The function uses list of `Color` generated from [PaletteGenerator] for calculations.
-  ///
-  /// **Note:**
-  /// - The values this function returns is computationally very expensive to calculate.
-  /// Consider using higher values for [skip] (which should not be more than the number of dominant colors in image) or
-  /// calculating luminance of the most dominant color generated from [PaletteGenerator] for an
-  /// image (use `computeLuminance()` of `Color`).
-  /// - For better accuracy of brightness/luminance, use [getBrightnessFrom] or[getWallpaperBrightness].
-  static Future<double> getLuminance(
-      {Uint8List imageData, int skip = 1}) async {
-    int index = 0, n = 0;
-    double luminance;
-    PaletteGenerator palette = await PaletteGenerator.fromUint8List(imageData);
-    double totalLum = 0;
-    while (index < palette.colors.length) {
-      totalLum += palette.colors.toList()[index].computeLuminance();
-      n += 1;
-      index += skip;
-    }
-    print('[getLuminance] Colors counted for calculations: $n');
-    luminance = totalLum / n;
-    return luminance;
-  }
-
   /// This gets the brightness of any image (image as `Uint8List`). The function returns
   /// a brightness level between 0 and 255, where 0 = totally black and 255 = totally bright.
   ///
@@ -129,7 +87,9 @@ class LauncherHelper {
   ///
   /// __Note:__
   /// - This method needs the READ_EXTERNAL_STORAGE permission on Android Oreo & above.
-  static getBrightnessFrom(Uint8List imageData, {int skipPixel = 1}) async {
+  /// - This uses kotlin implementation for calculating brightness.
+  static Future<int> calculateBrightness(Uint8List imageData,
+      {int skipPixel = 1}) async {
     assert(skipPixel > 0, 'skipPixel should have a value greater than 0');
     assert(imageData != null, 'imageData should not be null');
     int data = await _channel.invokeMethod(
@@ -137,35 +97,9 @@ class LauncherHelper {
     return data;
   }
 
-  /// A faster dart implementation to calculate brightness of an image. To replace [getLuminance].
-  /// This gets the brightness of any image (image as `Uint8List`). The function returns
-  /// a brightness level between 0 and 255, where 0 = totally black and 255 = totally bright.
-  ///
-  /// `skipPixel` parameter refers to number of pixels to skip while calculating Wallpaper's brightness.
-  /// `skipPixel` defaults to 1 (every pixel is counted) and can't be less than 1.
-  ///
-  /// __Note:__
-  /// - Still testing, I haven't completed anything in this. It's just an experimental implementation
-  /// for now, as I don't get much time to study on all this due to school. Recommended approach for now is [getBrightnessFrom]
-  static int calculateBrightness(Uint8List imageData, {int skipPixel = 1}) {
-    assert(skipPixel > 0, 'skipPixel should have a value greater than 0');
-    assert(imageData != null, 'imageData should not be null');
-    int count = 0, n = 0;
-    for (int pixel in imageData) {
-      count += pixel;
-      n += skipPixel;
-    }
-
-    return (count / (n * 3)).round();
-  }
-
-  /// It generates a palette based current Wallpaper to for use in UI colors.
-  ///
-  /// __Note:__
-  /// - This method needs the READ_EXTERNAL_STORAGE permission on Android Oreo & above.
-  static Future get wallpaperPalette async {
+  /// It returns a [PaletteGenerator] based on image (preferably wallpaper) to for use in UI colors.
+  static Future<PaletteGenerator> generatePalette(Uint8List imageData) async {
     PaletteGenerator _palette;
-    Uint8List imageData = await getWallpaper;
     _palette = await PaletteGenerator.fromUint8List(imageData);
     return _palette;
   }